[ESWE-1344] Prisoner base location (fix response, relocate fake gateway)
