+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsintegrationapi/controllers/v1/person/PrisonerBaseLocationController.kt
@@ -15,8 +15,8 @@ import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.config.FeatureFlagConfig
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.exception.EntityNotFoundException
+import uk.gov.justice.digital.hmpps.hmppsintegrationapi.models.hmpps.DataResponse
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.models.hmpps.PrisonerBaseLocation
-import uk.gov.justice.digital.hmpps.hmppsintegrationapi.models.hmpps.Response
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.models.hmpps.UpstreamApiError
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.models.roleconfig.ConsumerFilters
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.services.GetPrisonerBaseLocationForPersonService
@@ -42,19 +42,17 @@ class PrisonerBaseLocationController(
   fun getPrisonerBaseLocation(
     @Parameter(description = "A HMPPS id", example = "A123123") @PathVariable hmppsId: String,
     @RequestAttribute filters: ConsumerFilters?,
-  ): Response<PrisonerBaseLocation?> {
+  ): DataResponse<PrisonerBaseLocation?> {
     featureFlag.require(FeatureFlagConfig.USE_PRISONER_BASE_LOCATION_ENDPOINT)
 
     val response = getPrisonerBaseLocationForPersonService.execute(hmppsId, filters)
-    if (response.hasError(UpstreamApiError.Type.ENTITY_NOT_FOUND)) {
-      throw EntityNotFoundException("Could not find prisoner base location for id: $hmppsId")
-    }
-    if (response.hasError(UpstreamApiError.Type.BAD_REQUEST)) {
-      throw ValidationException("Invalid HMPPS ID: $hmppsId")
+    when {
+      response.hasError(UpstreamApiError.Type.ENTITY_NOT_FOUND) -> throw EntityNotFoundException("Could not find prisoner base location for id: $hmppsId")
+      response.hasError(UpstreamApiError.Type.BAD_REQUEST) -> throw ValidationException("Invalid HMPPS ID: $hmppsId")
     }
 
     auditService.createEvent("GET_PERSON_PRISONER_BASE_LOCATION", mapOf("hmppsId" to hmppsId))
 
-    return response
+    return DataResponse(data = response.data)
   }
 }
